Remove alert chips and start ship items
I decided not to include all the chips for alerts and only show a single chip in the header who's style is based on if we have critical alerts.
The ship list item was also started here with title and custom rendered subtitle for a quick view of ship status.